pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "self-mcp-scraper"
version = "0.1.0"
description = "Self-hosted MCP server that routes web scraping tools through your own HTTP or SOCKS5 proxy"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "self-mcp-scraper contributors" }]
keywords = ["mcp", "model-context-protocol", "web-scraping", "proxy", "ai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "httpx-socks>=0.9.2",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
screenshot = ["playwright>=1.44.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
self-mcp-scraper = "self_mcp_scraper.server:main"

[project.urls]
Homepage = "https://github.com/self-mcp-scraper/self-mcp-scraper"
Issues = "https://github.com/self-mcp-scraper/self-mcp-scraper/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/self_mcp_scraper"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
